+++ src/components/client/pages/calendar.css
@@ -1,0 +1,60 @@
+.timefix{
+    width: 18%;
+    padding: 4px;
+    margin-bottom: 12px;
+}
+
+button.addevent.btn.btn-primary {
+    margin-left: 426px;
+    margin-top: -20px;
+}
+
+.schedule {
+    width: 98%;
+    padding-left: 59px;
+    margin-top: -8px;
+}
+
+.rbc-date-cell > a, .rbc-date-cell > a:active, .rbc-date-cell > a:visited {
+    color: inherit;
+    text-decoration: none;
+    font-weight: bold;
+}
+
+.rbc-toolbar .rbc-toolbar-label {
+    flex-grow: 1;
+    padding: 0 10px;
+    text-align: center;
+    font-weight: bold;
+}
+
+.rbc-toolbar button {
+    color: #373a3c;
+    display: inline-block;
+    margin: 0;
+    text-align: center;
+    vertical-align: middle;
+    background: none;
+    background-image: none;
+    border: 1px solid #ccc;
+    padding: .375rem 1rem;
+    border-radius: 4px;
+    line-height: normal;
+    white-space: nowrap;
+    font-weight: bold;
+}
+
+.rbc-header {
+    overflow: hidden;
+    flex: 1 0;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    padding: 0 3px;
+    text-align: center;
+    vertical-align: middle;
+    font-weight: bold;
+    font-size: 90%;
+    min-height: 0;
+    border-bottom: 1px solid #DDD;
+    background-color: #bce3f3;
+}
