+++ src/components/common/navbar/nav.css
@@ -6,3 +6,10 @@ div#navbarSupportedContent {
 img {
   cursor: pointer;
 }
+
+nav.navbar.navbar-expand-lg.navbar-light.bg-light {
+  padding: 0px;
+}
+body{
+  margin: 0px;
+}
